iot-smart-env/prototypes/proto2_edge_rest_mqtt_min/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
paho-mqtt==2.1.0
SQLAlchemy==2.0.35
pydantic==2.9.2
python-dotenv==1.0.1